Update header for cast_raw_bytes_to_int
int64_t cast_raw_bytes_to_int

+++ software/NRG_itho_wifi/src/IthoSystem.h
@@ -141,7 +141,7 @@ void IthoPWMcommand(uint16_t value, volatile uint16_t *ithoCurrentVal, bool *upd
 int quick_pow10(int n);
 std::string i2cbuf2string(const uint8_t *data, size_t len);
 int cast_to_signed_int(int val, int length);
-int32_t cast_raw_bytes_to_int(int32_t* valptr, int length, bool is_signed);
+int64_t cast_raw_bytes_to_int(int32_t* valptr, int length, bool is_signed);
 uint32_t get_divider_from_datatype(int8_t datatype);
 uint8_t get_length_from_datatype(int8_t datatype);
 bool get_signed_from_datatype(int8_t datatype);
